Refactor documentation generation and add README rendering functionality
