+++ docs/__init__.py
@@ -1,5 +1,7 @@
-from .__main__ import main as documentation
+from .documentation import main as documentation
+from .readme import main as readme
 
 __all__ = [
     "documentation",
+    "readme",
 ]
