+++ tests/test_cli.py
@@ -1,3 +1,5 @@
+import warnings
+
 import pytest
 from click.testing import CliRunner
 
@@ -6,14 +8,15 @@ from jinja2_pdoc.cli import cli, jinja2pdoc
 
 def test_cli_folder(tmp_path):
     runner = CliRunner()
-    result = runner.invoke(
-        cli,
-        [
-            "docs/*.jinja2",
-            "--output",
-            str(tmp_path),
-        ],
-    )
+    with warnings.catch_warnings(record=True, category=DeprecationWarning):
+        result = runner.invoke(
+            cli,
+            [
+                "docs/*.jinja2",
+                "--output",
+                str(tmp_path),
+            ],
+        )
     assert result.exit_code == 0
     assert "rendering" in result.output
     assert tmp_path.joinpath("README.md").is_file()
